moved Dtype parsing to utils
