+++ litti-inference-router/internal/dto/featureStoreRecord.go
@@ -7,3 +7,12 @@ type FeatureStoreRecord struct {
 	ValidFrom      int
 	ValidTo        int
 }
+
+type FeatureEntry struct {
+	Name         string
+	Version      string
+	DataType     string
+	DefaultValue string
+	FeatureGroup string
+	TtlSeconds   int
+}
