+++ litti-inference-router/internal/feature/featureStore.go
@@ -4,8 +4,11 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
+	"fmt"
+	"strings"
 
 	"com/litti/ml/litti-inference-router/internal/dto"
+	"com/litti/ml/litti-inference-router/internal/util"
 
 	"github.com/redis/go-redis/v9"
 )
@@ -35,4 +38,49 @@ func fetchRedisFeatureGroup(featureGroupKey string) (map[string]dto.FeatureStore
 		return nil, err
 	}
 	return records, nil
+}
+
+func FetchRawFeatureRows(featureGroupSet map[string]bool, req dto.BatchPredictionRequest) (map[string]map[string]dto.FeatureStoreRecord, error) {
+	featureGroupKeys := make(map[string]bool)
+	reqFeatureGroupKeysMap := make(map[string][]string)
+	for _, predictionReq := range req.PredictionRequests {
+		if reqFeatureGroupKeysMap[predictionReq.Id] == nil {
+			reqFeatureGroupKeysMap[predictionReq.Id] = []string{}
+		}
+		for featureGroupName, _ := range featureGroupSet {
+			featureGroup := featureGroupMap[featureGroupName]
+			prefix := make([]string, len(featureGroup.Dimensions))
+			suffix := make([]string, len(featureGroup.Dimensions))
+			for _, dim := range featureGroup.Dimensions {
+				prefix = append(prefix, dim)
+				suffix = append(suffix, fmt.Sprint(predictionReq.Inputs[dim]))
+			}
+			featureGroupKey := strings.Join(prefix, "#") + "-" + strings.Join(suffix, "#")
+			featureGroupKeys[featureGroupKey] = true
+			reqFeatureGroupKeysMap[predictionReq.Id] = append(reqFeatureGroupKeysMap[predictionReq.Id], featureGroupKey)
+		}
+	}
+	// key os feature group key
+	featureGroupsMap := make(map[string]map[string]dto.FeatureStoreRecord)
+	for featureGroupKey := range featureGroupKeys {
+		featureStoreRecords, err := FetchFeatureGroupRow(featureGroupKey)
+		if err != nil {
+			return nil, err
+		}
+		if featureStoreRecords != nil {
+			featureGroupsMap[featureGroupKey] = featureStoreRecords
+		}
+	}
+	// key is reqId
+	// value is map[string]FeatureStoreRecord -> each entry in map represents a feature
+	reqFeatureRowsMap := make(map[string]map[string]dto.FeatureStoreRecord)
+	for reqId, fgKeys := range reqFeatureGroupKeysMap {
+		if reqFeatureRowsMap[reqId] == nil {
+			reqFeatureRowsMap[reqId] = map[string]dto.FeatureStoreRecord{}
+		}
+		for _, fgKey := range fgKeys {
+			reqFeatureRowsMap[reqId] = util.MergeMaps(reqFeatureRowsMap[reqId], featureGroupsMap[fgKey])
+		}
+	}
+	return reqFeatureRowsMap, nil
 }
