+++ litti-inference-router/internal/model/modelRoutingMetadata.go
@@ -1,24 +1,17 @@
 package model
 
+import "com/litti/ml/litti-inference-router/internal/dto"
+
 type ModelDeploymentMetadata struct {
 	Id       string
 	Name     string
 	Version  string
 	Domain   string
 	Outputs  []ModelOutputEntry
-	Features []FeatureEntry
+	Features []dto.FeatureEntry
 }
 
 type ModelOutputEntry struct {
 	Name     string
 	DataType string
-}
-
-type FeatureEntry struct {
-	Name         string
-	Version      string
-	DataType     string
-	DefaultValue string
-	FeatureGroup string
-	TtlSeconds   int
 }
